DEMOS-2385: Broke out some of the queries
Into a file so that it's more readable
